# src/app/docs/swagger-audiocore.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Audio Core API",
    "description": "Advanced audio generation system. Powered by Suno AI and ChatGPT.",
    "version": "1.2.0"
  },
...
  "paths": {
...
    "/api/personas": {
      "get": {
        "summary": "List User Personas",
        "description": "Get list of user's personas (voice profiles). Personas can be used to maintain vocal consistency across generated songs.",
        "tags": [
          "Persona"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": { "type": "number" },
            "description": "Page number for pagination (default: 1)"
          }
        ],
        "responses": {
          "200": {
            "description": "Personas retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "personas": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/persona_info"
                      }
                    },
                    "total_results": {
                      "type": "number",
                      "description": "Total number of personas"
                    },
                    "current_page": {
                      "type": "number",
                      "description": "Current page number"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/persona/delete": {
      "delete": {
        "summary": "Delete Persona",
        "description": "Delete (trash) a persona. Can also restore a previously deleted persona by setting undo=true.",
        "tags": [
          "Persona"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "required": true,
            "schema": { "type": "string" },
            "description": "The persona ID to delete",
            "example": "ce93a372-48f8-4ff7-aaa9-d57f871b75dd"
          },
          {
            "in": "query",
            "name": "undo",
            "schema": { "type": "boolean" },
            "description": "Set to true to restore a previously deleted persona (default: false)"
          }
        ],
        "responses": {
          "200": {
            "description": "Persona deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "Whether the operation was successful"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required parameter: id"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/persona/create": {
      "post": {
        "summary": "Create Persona",
        "description": "Create a persona (voice profile) from a song clip. First call /api/persona/vox-stem to extract vocals and get the vox_audio_id, then use it here. Personas are NOT public by default.",
        "tags": [
          "Persona"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["clip_id", "name"],
                "properties": {
                  "clip_id": {
                    "type": "string",
                    "description": "ID of the original song clip",
                    "example": "c3776feb-6b58-4229-aeb6-e27c716c77fc"
                  },
                  "name": {
                    "type": "string",
                    "description": "Name for the persona",
                    "example": "My Voice"
                  },
                  "description": {
                    "type": "string",
                    "description": "Optional description for the persona",
                    "example": "My custom voice profile for pop songs"
                  },
                  "is_public": {
                    "type": "boolean",
                    "description": "Whether the persona should be public (default: false)",
                    "default": false,
                    "example": false
                  },
                  "vox_audio_id": {
                    "type": "string",
                    "description": "ID from /api/persona/vox-stem response. Required for proper persona creation.",
                    "example": "1c9b8d6f-02d9-4446-bb4f-c4b8b939ced1"
                  },
                  "vocal_start_s": {
                    "type": "number",
                    "description": "Start time in seconds for the vocal sample (default: 0)",
                    "default": 0,
                    "example": 0
                  },
                  "vocal_end_s": {
                    "type": "number",
                    "description": "End time in seconds for the vocal sample (default: 30)",
                    "default": 30,
                    "example": 30
                  },
                  "user_input_styles": {
                    "type": "string",
                    "description": "Optional style tags for the persona",
                    "example": "rock, alternative, male vocals"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Persona created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/persona_info"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/persona/vox-stem": {
      "post": {
        "summary": "Extract Vox Stem",
        "description": "Extract vox (vocal) stem from a song for persona creation. This must be called before creating a persona to get the vox_audio_id. The extraction process isolates the vocal track from the specified time range.",
        "tags": [
          "Persona"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["clip_id"],
                "properties": {
                  "clip_id": {
                    "type": "string",
                    "description": "ID of the song to extract vocals from",
                    "example": "c3776feb-6b58-4229-aeb6-e27c716c77fc"
                  },
                  "vocal_start_s": {
                    "type": "number",
                    "description": "Start time in seconds for vocal extraction (default: 0)",
                    "default": 0,
                    "example": 45
                  },
                  "vocal_end_s": {
                    "type": "number",
                    "description": "End time in seconds for vocal extraction (default: 30). Max 30 seconds range.",
                    "default": 30,
                    "example": 75
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Vox stem extracted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "vox_audio_id": {
                      "type": "string",
                      "description": "ID of the extracted vox stem. Use this in /api/persona/create."
                    },
                    "clip_id": {
                      "type": "string",
                      "description": "Original clip ID"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "persona_info": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the persona"
          },
          "name": {
            "type": "string",
            "description": "Name of the persona"
          },
          "description": {
            "type": "string",
            "description": "Description of the persona"
          },
          "image_url": {
            "type": "string",
            "description": "URL of the persona's image"
          },
          "root_clip_id": {
            "type": "string",
            "description": "ID of the original clip used to create the persona"
          },
          "vox_audio_id": {
            "type": "string",
            "description": "ID of the extracted vox stem used for this persona"
          },
          "user_display_name": {
            "type": "string",
            "description": "Display name of the persona owner"
          },
          "user_handle": {
            "type": "string",
            "description": "Handle of the persona owner"
          },
          "is_public": {
            "type": "boolean",
            "description": "Whether the persona is public"
          },
          "is_owned": {
            "type": "boolean",
            "description": "Whether the current user owns the persona"
          },
          "clip_count": {
            "type": "number",
            "description": "Number of clips created with this persona"
          },
          "upvote_count": {
            "type": "number",
            "description": "Number of upvotes for this persona"
          }
        }
      }
    }
  }
}